refactor: move `handleIdentifyOrAlias`
